+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `RpcClientPipe` that supports working with Unix pipes
+
 ### Changed
 - `RpcUrl.password` is now expected to be only string and default value is empty
   string. `RpcUrl.login_type` is now `PLAIN` to be valid for default password